Develop

+++ database/seeders/DistrictSeeder.php
@@ -5,11 +5,15 @@ namespace Soap\ThaiAddresses\Database\Seeders;
 use File;
 use Illuminate\Database\Seeder;
 use Soap\ThaiAddresses\Models\District;
+use Soap\ThaiAddresses\ThaiAddresses;
 
 class DistrictSeeder extends Seeder
 {
     public function run()
     {
+        if (! \Schema::hasTable((new District)->getTable())) {
+            throw new \Exception('District table does not exist.');
+        }
         District::query()->delete();
         if (File::exists(base_path('vendor/soap/thai-addresses/database/seeders/data/districts.json'))) {
             $json = File::get(base_path('vendor/soap/thai-addresses/database/seeders/data/districts.json'));
@@ -24,7 +28,7 @@ class DistrictSeeder extends Seeder
                 'code' => $item->code,
                 'name_th' => $item->name_th,
                 'name_en' => $item->name_en,
-                config('thai-addresses.province.foreign_key') => $item->province_id,
+                ThaiAddresses::getProvinceForeignKeyName() => $item->province_id,
             ]);
         }
     }
