+++ database/seeders/ProvinceSeeder.php
@@ -4,13 +4,16 @@ namespace Soap\ThaiAddresses\Database\Seeders;
 
 use File;
 use Illuminate\Database\Seeder;
-use Soap\ThaiAddresses\Facades\ThaiAddresses;
 use Soap\ThaiAddresses\Models\Province;
+use Soap\ThaiAddresses\ThaiAddresses;
 
 class ProvinceSeeder extends Seeder
 {
     public function run()
     {
+        if (! \Schema::hasTable((new Province)->getTable())) {
+            throw new \Exception('Province table does not exist.');
+        }
         Province::query()->delete();
         if (File::exists(base_path('/vendor/soap/thai-addresses/database/seeders/data/provinces.json'))) {
             $json = File::get(base_path('/vendor/soap/thai-addresses/database/seeders/data/provinces.json'));
